refactor: deprecate about v1 api
